[Claude Code] set APNS_ENABLED=true set APNS_TEAM_ID=CR7GXH4RAJ set APNS_KEY_ID=45MQUUWGMX set prompt #246

+++ server/src/main/java/com/reminder/service/ReminderNotificationService.java
@@ -26,7 +26,7 @@ public class ReminderNotificationService {
         if (Boolean.FALSE.equals(reminder.getIosRepeatActive()) && reminder.getIosInitialPushedAt() != null) {
             return;
         }
-        int supporterCount = supporterRepository.countByReminderId(reminder.getId());
+        long supporterCount = supporterRepository.countByReminderId(reminder.getId());
         String timeStr = formatTime(reminder.getScheduledAt());
         String body;
         String sound;
